pyproject.toml
[project]
name = "dartwork-vega"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Sangwon Lee", email = "lsw91.main@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "altair>=5.5.0",
    "streamlit>=1.41.1",
    "vega-datasets>=0.9.0",
    "vl-convert-python>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
